+++ core/utils/scheduler.py
@@ -15,9 +15,7 @@ def run_crawler():
     logger.LoggerFactory.logbot.info("스케줄러에 의해 크롤러가 시작됩니다.")
     try:
         crawl_control.start_crawl(
-            crawl_mode=app_settings.crawl_mode,
-            crawl_type=app_settings.crawl_type,
-            max_empty_pages=app_settings.max_empty_pages,
+            crawl_mode="full",
             header="=== 자동 스케줄러 크롤링 작업 시작 ===",
             broadcast_source="scheduler",
         )